+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 - Added support for PrintingEnabled policy to enable or disable printing, see
   [#257](https://github.com/cadeyrn/enterprise-policy-generator/issues/257)
+- Added support for FirefoxSuggest policy to configure Firefox Suggest feature (US only), see
+  [#256](https://github.com/cadeyrn/enterprise-policy-generator/issues/256)
 - Added default_area option to ExtensionSettings policy to define the initial placement of the extension button, see
   [#255](https://github.com/cadeyrn/enterprise-policy-generator/issues/255)
 - Added EmailTracking option to EnableTrackingProtection policy to block e-mail tracking, see